% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CrownFractionBurned.r
\name{CrownFractionBurned}
\alias{CrownFractionBurned}
\title{Crown Fraction Burned Calculator}
\usage{
CrownFractionBurned(FUELTYPE, ROS, RSO)
}
\arguments{
\item{ROS}{Rate of Spread}

\item{RSO}{Critical Surface Rate of Spread}
}
\value{
CFB
}
\description{
Calculate Calculate Crown Fraction Burned.
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada,
  Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
