% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FireIntensity.r
\name{FireIntensity}
\alias{FireIntensity}
\title{Fire Intensity Calculator}
\usage{
FireIntensity(FC, ROS)
}
\arguments{
\item{FC}{Fuel Consumption (kg/m^2)}

\item{ROS}{Rate of Spread (m/min)}
}
\value{
FI: Fire Intensity (kW/m)
}
\description{
Calculate the Predicted Fire Intensity
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada,
  Ottawa, Ontario.
}
