% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\name{fireSeason}
\alias{fireSeason}
\title{Fire Season Start and End}
\usage{
fireSeason(
  input,
  fs.start = 12,
  fs.end = 5,
  method = "WF93",
  consistent.snow = FALSE,
  multi.year = FALSE
)
}
\description{
See \code{\link[cffdrs.core:fireSeason]{cffdrs.core::fireSeason()}}
}
