% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_FoliarMoistureContent.r
\name{FoliarMoistureContent}
\alias{FoliarMoistureContent}
\title{Foliar Moisture Content Calculator}
\usage{
FoliarMoistureContent(FUELTYPE, FMC, LAT, LONG, ELV, DJ, D0)
}
\arguments{
\item{LAT}{Latitude (decimal degrees)}

\item{LONG}{Longitude (decimal degrees)}

\item{ELV}{Elevation (metres)}

\item{DJ}{Day of year (offeren referred to as julian date)}

\item{D0}{Date of minimum foliar moisture content. _If D0, date of min FMC, is not known then D0 = NULL._}
}
\value{
FMC: Foliar Moisture Content value
}
\description{
Calculate Foliar Moisture Content on a specified day.
All variables names are laid out in the same manner as Forestry Canada
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.
}
