% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Fuels.r
\docType{data}
\name{FUELS}
\alias{FUELS}
\title{\tabular{ll}{
\bold{Fuel Type} \tab \bold{code} \cr 
\verb{C-1}       \tab 1           \cr 
\verb{C-2}       \tab 2           \cr 
\verb{C-3}       \tab 3           \cr
\verb{C-4}       \tab 4           \cr 
\verb{C-5}       \tab 5           \cr 
\verb{C-6}       \tab 6           \cr 
\verb{C-7}       \tab 7           \cr 
\verb{D-1}       \tab 8           \cr 
\verb{M-1}       \tab 9           \cr 
\verb{M-2}       \tab 10          \cr 
\verb{M-3}       \tab 11          \cr 
\verb{M-4}       \tab 12          \cr 
\verb{NF}        \tab 13          \cr 
\verb{O-1a}      \tab 14          \cr 
\verb{O-1b}      \tab 15          \cr 
\verb{S-1}       \tab 16          \cr 
\verb{S-2}       \tab 17          \cr 
\verb{S-3}       \tab 18          \cr 
\verb{WA}        \tab 19          \cr\cr}}
\format{
An object of class \code{list} of length 19.
}
\usage{
FUELS
}
\description{
\tabular{ll}{
\bold{Fuel Type} \tab \bold{code} \cr 
\verb{C-1}       \tab 1           \cr 
\verb{C-2}       \tab 2           \cr 
\verb{C-3}       \tab 3           \cr
\verb{C-4}       \tab 4           \cr 
\verb{C-5}       \tab 5           \cr 
\verb{C-6}       \tab 6           \cr 
\verb{C-7}       \tab 7           \cr 
\verb{D-1}       \tab 8           \cr 
\verb{M-1}       \tab 9           \cr 
\verb{M-2}       \tab 10          \cr 
\verb{M-3}       \tab 11          \cr 
\verb{M-4}       \tab 12          \cr 
\verb{NF}        \tab 13          \cr 
\verb{O-1a}      \tab 14          \cr 
\verb{O-1b}      \tab 15          \cr 
\verb{S-1}       \tab 16          \cr 
\verb{S-2}       \tab 17          \cr 
\verb{S-3}       \tab 18          \cr 
\verb{WA}        \tab 19          \cr\cr}
}
\keyword{datasets}
