% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\name{fwi}
\alias{fwi}
\title{Fire Weather Index System}
\usage{
fwi(
  input,
  init = data.frame(ffmc = 85, dmc = 6, dc = 15, lat = 55),
  batch = TRUE,
  out = "all",
  lat.adjust = TRUE,
  uppercase = TRUE
)
}
\description{
See \code{\link[cffdrs.core:fwi]{cffdrs.core::fwi()}}
}
