% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\name{gfmc}
\alias{gfmc}
\title{Grass Fuel Moisture Code}
\usage{
gfmc(
  input,
  GFMCold = 85,
  batch = TRUE,
  time.step = 1,
  roFL = 0.3,
  out = "GFMCandMC"
)
}
\description{
See \code{\link[cffdrs.core:gfmc]{cffdrs.core::gfmc()}}
}
