% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\name{hffmc}
\alias{hffmc}
\title{Hourly Fine Fuel Moisture Code}
\usage{
hffmc(
  weatherstream,
  ffmc_old = 85,
  time.step = 1,
  calc.step = FALSE,
  batch = TRUE,
  hourlyFWI = FALSE
)
}
\description{
See \code{\link[cffdrs.core:hffmc]{cffdrs.core::hffmc()}}
}
