% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_HourlyFineFuelMoistureCode.r
\name{HourlyFineFuelMoistureCode}
\alias{HourlyFineFuelMoistureCode}
\title{Hourly Fine Fuel Moisture Code Calculation}
\usage{
HourlyFineFuelMoistureCode(temp, rh, ws, prec, ffmc_old, time.step = 1)
}
\arguments{
\item{temp}{Temperature (centigrade)}

\item{rh}{Relative Humidity (\%)}

\item{ws}{Wind speed (km/h)}

\item{prec}{Precipitation (mm)}

\item{ffmc_old}{The Fine Fuel Moisture Code from previous iteration}

\item{time.step}{Time step (hours) [default 1 hour]}
}
\value{
A single hourly fine fuel moisture code value
}
\description{
Hourly Fine Fuel Moisture Code Calculation
}
