% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_RateOfSpreadAtTheta.r
\name{RateOfSpreadAtTheta}
\alias{RateOfSpreadAtTheta}
\title{Rate of spread at a point along the perimeter calculator}
\usage{
RateOfSpreadAtTheta(ROS, FROS, BROS, THETA)
}
\arguments{
\item{ROS}{Rate of Spread (m/min)}

\item{FROS}{Flank Fire Rate of Spread (m/min)}

\item{BROS}{Back Fire Rate of Spread (m/min)}

\item{THETA}{}

\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}
}
\value{
ROSTHETA - Rate of spread at point theta(m/min)
}
\description{
Computes the Rate of Spread at any point along the perimeter of an 
elliptically shaped fire. Equations are from Wotton et. al. (2009).
}
\details{
Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
Canada. Information Report GLC-X-10, 45p.
}
