\name{shortwave_model}
\docType{data}
\alias{shortwave_model}
\title{Fire Weather Index Sample Input Data Set}
\description{
  This data set is the sample input data that was used in original FWI program calibration.
}
\usage{data(shortwave_model)}
\format{A data frame containing 9 columns and 49 rows, with 1 header line}
\source{
\url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/19973.pdf}
}
\references{
1.	Van Wagner, CE. and T.L. Pickett. 1985. Equations and FORTRAN program for the Canadian Forest Fire Weather Index System. Can. For. Serv., Ottawa, Ont. For. Tech. Rep. 33. 18 p.
}
\keyword{datasets}
