% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SurfaceRateOfSpreadC6.r
\name{SurfaceRateOfSpreadC6}
\alias{SurfaceRateOfSpreadC6}
\title{C-6 Conifer Plantation Fire Spread Calculator}
\usage{
SurfaceRateOfSpreadC6(RSI, BUI)
}
\arguments{
\item{RSI}{Intermediate Surface Fire Rate of Spread}

\item{BUI}{Buildup Index}

\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}
}
\value{
RSS
}
\description{
Calculate surface fire rate of spread (RSS).
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada, 
 Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
