% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_fbp}
\alias{test_fbp}
\title{See \code{\link[cffdrs.core:test_fbp]{cffdrs.core::test_fbp}}}
\format{
An object of class \code{data.frame} with 20 rows and 24 columns.
}
\usage{
test_fbp
}
\description{
See \code{\link[cffdrs.core:test_fbp]{cffdrs.core::test_fbp}}
}
\keyword{datasets}
