% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_fwi}
\alias{test_fwi}
\title{See \code{\link[cffdrs.core:test_fwi]{cffdrs.core::test_fwi}}}
\format{
An object of class \code{data.frame} with 48 rows and 9 columns.
}
\usage{
test_fwi
}
\description{
See \code{\link[cffdrs.core:test_fwi]{cffdrs.core::test_fwi}}
}
\keyword{datasets}
