% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_gfmc}
\alias{test_gfmc}
\title{See \code{\link[cffdrs.core:test_gfmc]{cffdrs.core::test_gfmc}}}
\format{
An object of class \code{data.frame} with 198 rows and 9 columns.
}
\usage{
test_gfmc
}
\description{
See \code{\link[cffdrs.core:test_gfmc]{cffdrs.core::test_gfmc}}
}
\keyword{datasets}
