% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_hffmc}
\alias{test_hffmc}
\title{See \code{\link[cffdrs.core:test_hffmc]{cffdrs.core::test_hffmc}}}
\format{
An object of class \code{data.frame} with 480 rows and 8 columns.
}
\usage{
test_hffmc
}
\description{
See \code{\link[cffdrs.core:test_hffmc]{cffdrs.core::test_hffmc}}
}
\keyword{datasets}
