% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_sdmc}
\alias{test_sdmc}
\title{See \code{\link[cffdrs.core:test_sdmc]{cffdrs.core::test_sdmc}}}
\format{
An object of class \code{data.frame} with 48 rows and 10 columns.
}
\usage{
test_sdmc
}
\description{
See \code{\link[cffdrs.core:test_sdmc]{cffdrs.core::test_sdmc}}
}
\keyword{datasets}
