% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_wDC}
\alias{test_wDC}
\title{See \code{\link[cffdrs.core:test_wDC]{cffdrs.core::test_wDC}}}
\format{
An object of class \code{data.frame} with 2192 rows and 11 columns.
}
\usage{
test_wDC
}
\description{
See \code{\link[cffdrs.core:test_wDC]{cffdrs.core::test_wDC}}
}
\keyword{datasets}
