% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposed.r
\docType{data}
\name{test_wDC_fs}
\alias{test_wDC_fs}
\title{See \code{\link[cffdrs.core:test_wDC_fs]{cffdrs.core::test_wDC_fs}}}
\format{
An object of class \code{data.frame} with 8 rows and 7 columns.
}
\usage{
test_wDC_fs
}
\description{
See \code{\link[cffdrs.core:test_wDC_fs]{cffdrs.core::test_wDC_fs}}
}
\keyword{datasets}
